requirements.txt
google-genai>=1.0.0
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
supabase>=2.5.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
openai-whisper>=20231117
soundfile>=0.12.1
kokoro-onnx>=0.1.0
twilio>=9.0.0
